requirements.txt
requests
beautifulsoup4
python-dotenv
selenium 
feedparser
load_dotenv
